requirements.txt
pydantic==2.5.3
PyMuPDF==1.27.1

llama-index-core
torch
transformers
sentence-transformers
qdrant-client>=1.7.1,<1.16
llama-index-vector-stores-qdrant
llama-index-llms-ollama
llama-index-embeddings-ollama
uvicorn[standard]
